+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 - Remove CAPA templating from `aws` provider.
 - Add new provider `capa` for templating a cluster.
 - Remove fetching ssh sso ca pub key for capa from management cluster.
+- Add test for CAPA provider cluster templating.
 
 ## [2.18.0] - 2022-07-08